% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_backward_linkages}
\alias{compute_backward_linkages}
\title{Computes backward linkages}
\usage{
compute_backward_linkages(leontief_inverse_matrix)
}
\arguments{
\item{leontief_inverse_matrix}{A nxn matrix of Leontief inverse.}
}
\value{
A vector of backward linkages.
}
\description{
Computes backward linkages from a Leontief inverse matrix.
}
