% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_extraction_backward}
\alias{compute_extraction_backward}
\title{Calculates backward linkage extraction.}
\usage{
compute_extraction_backward(
  technical_coefficients_matrix,
  final_demand_matrix,
  total_production
)
}
\arguments{
\item{technical_coefficients_matrix}{A nxn matrix of technical coefficients.}

\item{final_demand_matrix}{The final demand matrix.}

\item{total_production}{A 1xn vector of total production.}
}
\description{
Computes impact on demand structure after extracting a given sector.
}
\references{
\cite{Vale, Vinícius de Almeida, e Fernando Salgueiro Perobelli. Análise de Insumo-Produto: teoria e aplicações no R. Curitiba, PR: Edição Independente, 2020.}
}
