% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_extraction_total}
\alias{compute_extraction_total}
\title{Calculates total extraction}
\usage{
compute_extraction_total(backward_linkage_matrix, forward_linkage_matrix)
}
\arguments{
\item{backward_linkage_matrix}{A nx2 matrix of backward linkage.}

\item{forward_linkage_matrix}{A nx2 matrix of forward linkage.}
}
\description{
Computes total impact after extracting a given sector.
}
