% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_field_influence}
\alias{compute_field_influence}
\title{Calculates field of influence given a technical change.}
\usage{
compute_field_influence(tech_coeff_matrix, leontief_inverse_matrix, epsilon)
}
\arguments{
\item{tech_coeff_matrix}{A nxn matrix of technical coefficients.}

\item{leontief_inverse_matrix}{The open model nxn Leontief inverse matrix.}

\item{epsilon}{The epsilon value.}
}
\value{
Field of influence matrix.
}
\description{
Calculates total field of influence given a incremental change in the techincal coefficients matrix.
}
