% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_forward_linkages}
\alias{compute_forward_linkages}
\title{Computes foward linkages}
\usage{
compute_forward_linkages(leontief_inverse_matrix)
}
\arguments{
\item{leontief_inverse_matrix}{A nxn matrix of Leontief inverse.}
}
\value{
A vector of forward linkages.
}
\description{
Computes forward linkages from a Leontief inverse matrix.
}
