% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_generator_added_value}
\alias{compute_generator_added_value}
\title{Calculates generator matrix for a given added value vector.}
\usage{
compute_generator_added_value(
  added_value_requirements,
  leontief_inverse_matrix
)
}
\arguments{
\item{added_value_requirements}{The coefficients for a given added value vector.}

\item{leontief_inverse_matrix}{The open model Leontief inverse matrix.}
}
\value{
A nxn matrix of an added value vector generator.
}
\description{
Calculates generator matrix for a given added value vector.
}
