% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_generator_employment}
\alias{compute_generator_employment}
\title{Calculates employment generator matrix.}
\usage{
compute_generator_employment(employment_requirements, leontief_inverse_matrix)
}
\arguments{
\item{employment_requirements}{The employment requirements vector.}

\item{leontief_inverse_matrix}{The open model Leontief inverse matrix.}
}
\value{
A nxn matrix of employment generator.
}
\description{
Calculates employment generator matrix.
}
