% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_ghosh_inverse}
\alias{compute_ghosh_inverse}
\title{Computes Ghosh inverse matrix.}
\usage{
compute_ghosh_inverse(allocation_coeff)
}
\arguments{
\item{allocation_coeff}{A nxn matrix of allocation coefficients.}
}
\value{
A nxn matrix of Ghosh inverse.
}
\description{
Computes Ghosh inverse matrix.
}
