% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_leontief_inverse}
\alias{compute_leontief_inverse}
\title{Computes Leontief inverse matrix to R.}
\usage{
compute_leontief_inverse(tech_coeff)
}
\arguments{
\item{tech_coeff}{A nxn matrix of technical coefficients.}
}
\value{
A nxn matrix of Leontief inverse.
}
\description{
Computes Leontief inverse matrix to R.
}
