% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_multiplier_added_value}
\alias{compute_multiplier_added_value}
\title{Calculates multiplier for a given added value vector.}
\usage{
compute_multiplier_added_value(
  added_value_requirements,
  leontief_inverse_matrix
)
}
\arguments{
\item{added_value_requirements}{The coefficients for a given added value vector.}

\item{leontief_inverse_matrix}{The open model Leontief inverse matrix.}
}
\description{
Calculates multiplier for a given added value vector.
}
