% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_multiplier_added_value_indirect}
\alias{compute_multiplier_added_value_indirect}
\title{Computes indirect multiplier for a given added value vector.}
\usage{
compute_multiplier_added_value_indirect(
  added_value_element,
  total_production,
  leontief_inverse_matrix
)
}
\arguments{
\item{added_value_element}{An added value vector.}

\item{total_production}{The total production vector.}

\item{leontief_inverse_matrix}{The open model Leontief inverse matrix.}
}
\value{
A 1xn vector of indirect multipliers for a given added value vector.
}
\description{
Computes indirect multiplier for a given added value vector.
}
