% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_multiplier_employment}
\alias{compute_multiplier_employment}
\title{Calculates type I employment multiplier.}
\usage{
compute_multiplier_employment(employment_requirements, leontief_inverse_matrix)
}
\arguments{
\item{employment_requirements}{The employment requirements.}

\item{leontief_inverse_matrix}{The open model Leontief inverse matrix.}
}
\description{
Calculates type I employment multiplier.
}
