% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_multiplier_employment_indirect}
\alias{compute_multiplier_employment_indirect}
\title{Calculates type I indirect employment multiplier.}
\usage{
compute_multiplier_employment_indirect(
  employment_levels,
  total_production,
  leontief_inverse_matrix
)
}
\arguments{
\item{leontief_inverse_matrix}{The open model Leontief inverse matrix.}

\item{employment_requirements}{The employment requirements.}
}
\value{
A 1xn vector of type I indirect employment multipliers.
}
\description{
Calculates type I indirect employment multiplier.
}
