% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_multiplier_output}
\alias{compute_multiplier_output}
\title{\itemize{
\item MARK: output multipliers
Calculates type I output multiplier.
}}
\usage{
compute_multiplier_output(leontief_inverse_matrix)
}
\arguments{
\item{leontief_inverse_matrix}{The open model Leontief inverse matrix.}
}
\value{
A 1xn vector of type I output multipliers.
}
\description{
\itemize{
\item MARK: output multipliers
Calculates type I output multiplier.
}
}
