% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_multiplier_output_direct}
\alias{compute_multiplier_output_direct}
\title{Computes direct output multiplier.}
\usage{
compute_multiplier_output_direct(technical_coefficients_matrix)
}
\arguments{
\item{technical_coefficients_matrix}{The open model technical coefficients matrix.}
}
\value{
A 1xn vector of direct output multipliers.
}
\description{
Computes direct output multiplier.
}
