% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_multiplier_output_indirect}
\alias{compute_multiplier_output_indirect}
\title{Calculates type I indirect output multiplier.}
\usage{
compute_multiplier_output_indirect(
  technical_coefficients_matrix,
  leontief_inverse_matrix
)
}
\arguments{
\item{technical_coefficients_matrix}{The open model technical coefficients matrix.}

\item{leontief_inverse_matrix}{The open model Leontief inverse matrix.}
}
\value{
A 1xn vector of type I indirect output multipliers.
}
\description{
Calculates type I indirect output multiplier.
}
