% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_power_dispersion}
\alias{compute_power_dispersion}
\title{Computes power of dispersion}
\usage{
compute_power_dispersion(leontief_inverse_matrix)
}
\arguments{
\item{leontief_inverse_matrix}{A nxn matrix of Leontief inverse.}
}
\value{
A vector of power of dispersion.
}
\description{
Computes power of dispersion from a Leontief inverse matrix.
}
