% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_requirements_added_value}
\alias{compute_requirements_added_value}
\title{Calculates requirements for a given added value vector}
\usage{
compute_requirements_added_value(added_value_element, total_production)
}
\arguments{
\item{added_value_element}{An added value vector.}

\item{total_production}{The total production vector.}
}
\value{
A 1xn vector of a given added value coefficients.
}
\description{
Calculates requirements for a given added value vector
}
