% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_requirements_employment}
\alias{compute_requirements_employment}
\title{\itemize{
\item MARK: Employment Multipliers
Calculates employment requirements.
}}
\usage{
compute_requirements_employment(employment_levels, total_production)
}
\arguments{
\item{employment_levels}{The employment levels.}

\item{total_production}{The total production.}
}
\value{
A 1xn vector of employment requirements.
}
\description{
\itemize{
\item MARK: Employment Multipliers
Calculates employment requirements.
}
}
