% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_sensitivity_dispersion}
\alias{compute_sensitivity_dispersion}
\title{Computes sensitivity of dispersion}
\usage{
compute_sensitivity_dispersion(leontief_inverse_matrix)
}
\arguments{
\item{leontief_inverse_matrix}{A nxn matrix of Leontief inverse.}
}
\value{
A vector of sensitivity of dispersion.
}
\description{
Computes sensitivity of dispersion from a Leontief inverse matrix.
}
