% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_sensitivity_dispersion_cv}
\alias{compute_sensitivity_dispersion_cv}
\title{Computes sensitivity of dispersion coefficients of variation}
\usage{
compute_sensitivity_dispersion_cv(leontief_inverse_matrix)
}
\arguments{
\item{leontief_inverse_matrix}{A nxn matrix of Leontief inverse.}
}
\value{
A vector of sensitivity of dispersion coefficients of variation.
}
\description{
Computes sensitivity of dispersion coefficients of variation of an economy.
}
