% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_tech_coeff}
\alias{compute_tech_coeff}
\title{Computes technical coefficients matrix to R.}
\usage{
compute_tech_coeff(intermediate_transactions, total_production)
}
\arguments{
\item{intermediate_transactions}{A nxn matrix of intermediate transactions.}

\item{total_production}{A 1xn vector of total production.}
}
\value{
A nxn matrix of technical coefficients, known as A matrix.
}
\description{
Computes technical coefficients matrix to R.
}
