% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{extraction_backward}
\alias{extraction_backward}
\title{Calculates backward linkage extraction.}
\usage{
extraction_backward(tech_coeff_matrix, total_production, final_demand_matrix)
}
\arguments{
\item{tech_coeff_matrix}{A nxn matrix of technical coefficients.}

\item{total_production}{A 1xn vector of total production.}

\item{final_demand_matrix}{The final demanda matrix.}
}
\description{
Computes impact on demand structure after extracting a given sector.
}
