% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{extraction_forward}
\alias{extraction_forward}
\title{Calculates forward linkage extraction.}
\usage{
extraction_forward(
  allocation_coefficients_matrix,
  total_production,
  added_value_matrix
)
}
\arguments{
\item{total_production}{A 1xn vector of total production.}

\item{added_value_matrix}{The added value matrix.}

\item{allocation_coeff_matrix}{A nxn matrix of allocation coefficients.}
}
\description{
Computes impact on supply structure after extracting a given sector.
}
