% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{fio_addin}
\alias{fio_addin}
\title{Conveniently import data from an Excel file}
\usage{
fio_addin()
}
\description{
\code{fio_addin()} opens an \href{https://shiny.rstudio.com/articles/gadgets.html}{RStudio gadget} and
\href{https://rstudio.github.io/rstudioaddins/}{addin} that allows you to say
where the data source is (either clipboard or Excel file) and import the
data into the global environment.
Appears as "Import input-output data" in the RStudio Addins menu.
}
