% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\docType{data}
\name{input_options}
\alias{input_options}
\title{Conveniently import data from an Excel file}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
input_options
}
\description{
\code{fio_addin()} opens an \href{https://shiny.rstudio.com/articles/gadgets.html}{RStudio gadget} and
\href{https://rstudio.github.io/rstudioaddins/}{addin} that allows you to say
where the data source is (either clipboard or Excel file) and import the
data into the global environment.
Appears as "Import input-output data" in the RStudio Addins menu.
}
\keyword{datasets}
