% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{leontief_inverse}
\alias{leontief_inverse}
\title{Calculates Leontief inverse matrix to R.}
\usage{
leontief_inverse(tec_coeff)
}
\arguments{
\item{tec_coeff}{A nxn matrix of technical coefficients.}
}
\value{
A nxn matrix of Leontief inverse.
}
\description{
Calculates Leontief inverse matrix to R.
}
\examples{
tec_coeff <- matrix(c(0.1, 0.2, 0.3, 0.4), nrow = 2)
leontief_inverse(tec_coeff)
}
