% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{set_max_threads}
\alias{set_max_threads}
\title{Sets max number of threads used by fio}
\usage{
set_max_threads(max_threads)
}
\arguments{
\item{max_threads}{Max number of threads enable globally for fio. 0 means all threads available.}
}
\value{
This functions does not return a value.
}
\description{
Sets max number of threads used by fio
}
\details{
Calling this function sets a global limit of threads to Rayon crate, affecting
all computations that runs in parallel by default.
}
\examples{
intermediate_transactions <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), 3, 3)
total_production <- matrix(c(100, 200, 300), 1, 3)
# instantiate iom object
my_iom <- fio::iom$new("test", intermediate_transactions, total_production)
# make Rust code run in sequence (disable parallelization)
my_iom$set_max_threads(1)
}
