% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{tec_coeff}
\alias{tec_coeff}
\title{Calculates technical coefficients matrix to R.}
\usage{
tec_coeff(intermediate_transactions, total_production)
}
\arguments{
\item{intermediate_transactions}{A nxn matrix of intermediate transactions.}

\item{total_production}{A 1xn vector of total production.}
}
\value{
A nxn matrix of technical coefficients, known as A matrix.
}
\description{
Calculates technical coefficients matrix to R.
}
\examples{
intermediate_transactions <- matrix(c(1, 2, 3, 4), nrow = 2)
total_production <- c(10, 20)
tec_coeff(intermediate_transactions, total_production)
}
