% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{tecnical_coef}
\alias{tecnical_coef}
\title{Import IOM data}
\usage{
tecnical_coef(data)
}
\arguments{
\item{file}{Path to the Excel file.}

\item{sheet}{Name of the sheet in the Excel file.}

\item{range}{Range of cells in the Excel file.}

\item{col_names}{Range of cells with column names.}

\item{row_names}{Range of cells with row names.}
}
\value{
A matrix with row and column names.
}
\description{
Import data from a input-output matrix (IOM) from Excel format.
}
\examples{
\dontrun{
z = import_element(
 file = "MIP-BR (2020).xlsx",
 sheet = "MIP",
 range = "D6:BB56",
 col_names = "D4:BB4",
 row_names = "D4:BB4"
)
}
}
