% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_allocation_coeff}
\alias{compute_allocation_coeff}
\title{Computes allocation coefficients matrix.}
\usage{
compute_allocation_coeff(intermediate_transactions, total_production)
}
\arguments{
\item{intermediate_transactions}{A nxn matrix of intermediate transactions.}

\item{total_production}{A 1xn vector of total production.}
}
\value{
A nxn matrix of allocation coefficients, known as F matrix.
}
\description{
Computes allocation coefficients matrix.
}
\details{
It calculates the allocation coefficients matrix, which is the rowwise ratio of
intermediate transactions to total production \insertCite{miller_input-output_2009}{fio}.

Underlined Rust code runs in parallel by default, so there is no need to
use future or async/await to parallelize.
}
\examples{
# running single threaded to comply with CRAN policies. Ignore for performance.
fio:::set_max_threads(1L)
intermediate_transactions <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), 3, 3)
total_production <- matrix(c(100, 200, 300), 1, 3)
# instantiate iom object
my_iom <- fio::iom$new("test", intermediate_transactions, total_production)
# Calculate the allocation coefficients
my_iom$compute_allocation_coeff()
# show the allocation coefficients
my_iom$allocation_coefficients_matrix

}
\references{
insertAllCited{}
}
