% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_extraction_forward}
\alias{compute_extraction_forward}
\title{Calculates forward linkage extraction.}
\usage{
compute_extraction_forward(
  allocation_coefficients_matrix,
  added_value_matrix,
  total_production
)
}
\arguments{
\item{allocation_coefficients_matrix}{A nxn matrix of allocation coefficients.}

\item{added_value_matrix}{The added value matrix.}

\item{total_production}{A 1xn vector of total production.}
}
\description{
Computes impact on supply structure after extracting a given sector.
}
\references{
\cite{Vale, Vinícius de Almeida, e Fernando Salgueiro Perobelli. Análise de Insumo-Produto: teoria e aplicações no R. Curitiba, PR: Edição Independente, 2020.}
}
