% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_extraction_total}
\alias{compute_extraction_total}
\title{Calculates total extraction}
\usage{
compute_extraction_total(backward_linkage_matrix, forward_linkage_matrix)
}
\arguments{
\item{backward_linkage_matrix}{A nx2 matrix of backward linkage.}

\item{forward_linkage_matrix}{A nx2 matrix of forward linkage.}
}
\description{
Computes total impact after extracting a given sector.
}
\details{
Here we define total impact as the sum of impact on demand and supply structures
after removal of a given sector.
}
\examples{
# running single threaded to comply with CRAN policies. Ignore for performance.
fio:::set_max_threads(1L)
intermediate_transactions <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), 3, 3)
total_production <- matrix(c(100, 200, 300), 1, 3)
exports <- matrix(c(10, 20, 30), 3, 1)
imports <- matrix(c(5, 10, 15), 1, 3)
# instantiate iom object
my_iom <- fio::iom$new(
  "test",
  intermediate_transactions,
  total_production,
  exports = exports,
  imports = imports
)

# Calculate the technical coefficients
my_iom$compute_tech_coeff()
# calculate the Leontief inverse
my_iom$compute_allocation_coeff()
# aggregate final demand and added value matrices
my_iom$update_added_value_matrix()
my_iom$update_final_demand_matrix()
# Calculate effects on both demand and supply structures after extracting a sector
my_iom$compute_hypothetical_extraction()
# show results
my_iom$hypothetical_extraction
}
\seealso{
\code{compute_extraction_backwards()} and \code{compute_extraction_forward()}.
}
