% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_field_influence}
\alias{compute_field_influence}
\title{Computes field of influence given a technical change.}
\usage{
compute_field_influence(tech_coeff_matrix, leontief_inverse_matrix, epsilon)
}
\arguments{
\item{tech_coeff_matrix}{A nxn matrix of technical coefficients.}

\item{leontief_inverse_matrix}{The open model nxn Leontief inverse matrix.}

\item{epsilon}{The epsilon value.}
}
\value{
Field of influence matrix.
}
\description{
Computes total field of influence given a incremental change in the technical coefficients matrix \insertCite{vale_alise_2020}{fio}.
}
\examples{
intermediate_transactions <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), 3, 3)
total_production <- matrix(c(100, 200, 300), 1, 3)
# instantiate iom object
my_iom <- fio::iom$new("test", intermediate_transactions, total_production)
# calculate the technical coefficients
my_iom$compute_tech_coeff()
# calculate the Leontief inverse
my_iom$compute_leontief_inverse()
# calculate field of influence
my_iom$compute_field_influence(epsilon = 0.01)
my_iom$field_influence
}
\references{
\insertAllCited{}
}
