% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_ghosh_inverse}
\alias{compute_ghosh_inverse}
\title{Computes Ghosh inverse matrix.}
\usage{
compute_ghosh_inverse(allocation_coeff)
}
\arguments{
\item{allocation_coeff}{A \eqn{n x n} matrix of allocation coefficients.}
}
\value{
A \eqn{n x n} matrix of Ghoshian inverse.
}
\description{
Computes Ghosh inverse matrix.
}
\details{
The Ghosh inverse matrix is the inverse of the
difference \eqn{(I - F)} where I is the identity matrix and F is the
allocation coefficients matrix \insertCite{miller_input-output_2009}{fio}.
}
\references{
\insertAllCited{}
}
