% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compute_ghosh_inverse}
\alias{compute_ghosh_inverse}
\title{Computes Ghosh inverse matrix.}
\usage{
compute_ghosh_inverse(allocation_coeff)
}
\arguments{
\item{allocation_coeff}{A \eqn{n x n} matrix of allocation coefficients.}
}
\value{
A \eqn{n x n} matrix of Ghoshian inverse.
}
\description{
Computes Ghosh inverse matrix.
}
\details{
It calculates the Ghosh inverse matrix, which is the inverse of the
difference \eqn{(I - F)} where I is the identity matrix and F is the
allocation coefficients matrix.
}
\examples{
intermediate_transactions <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), 3, 3)
total_production <- matrix(c(100, 200, 300), 1, 3)
# instantiate iom object
my_iom <- fio::iom$new("test", intermediate_transactions, total_production)
# disable parallelization for CRAN checks
my_iom$set_max_threads(1)
# Calculate the allocation coefficients
my_iom$compute_allocation_coeff()
# Calculate the Ghosh inverse
my_iom$compute_ghosh_inverse()
# show the Ghosh inverse
my_iom$ghosh_inverse_matrix
}
\references{
\cite{Miller, Ronald E., and Peter D. Blair. Input-Output Analysis: Foundations and Extensions. 2nd ed. Cambridge University Press, 2009. https://doi.org/10.1017/CBO9780511626982.}
}
