% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wiod.R
\name{download_wiod}
\alias{download_wiod}
\title{Download WIOD tables}
\usage{
download_wiod(year = "2016", out_dir = tempdir())
}
\arguments{
\item{year}{(\code{string})\cr
Release year from WIOD. One of "2016", "2013" or "long-run".
Defaults to "2016".}

\item{out_dir}{(\code{string})\cr
Path to download. Defaults to \code{tempdir()}.}
}
\value{
Invisibly returns the path to the downloaded file.
}
\description{
Downloads World Input-Output Database tables.
}
\details{
Multi-region input-output tables from the World Input-Output Database
(WIOD) from University of Groningen, Netherlands.
}
\examples{
\dontrun{
fio::download_wiod("2016")
}
}
