% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{download_wiod}
\alias{download_wiod}
\title{Download WIOD tables}
\usage{
download_wiod(year = "2016", out_dir = getwd())
}
\arguments{
\item{year}{(\code{string})\cr
Release year from WIOD. One of "2016", "2013" or "long-run". Defaults to "2016".}

\item{out_dir}{(\code{string})\cr
Path to download. Defaults to current working directory.}
}
\value{
A message indicating the result of the download operation.
}
\description{
Downloads World Input-Output Database tables.
}
\details{
Multi-region input-output tables from the World Input-Output Database (WIOD) from University of Groningen, Netherlands.
}
\examples{
# Download WIOD 2016 tables to temporary directory
\dontrun{
  fio::download_wiod("2016", getwd())
}
}
