% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{fio-package}
\alias{fio}
\alias{fio-package}
\title{fio: Friendly Input-Output Analysis}
\description{
A set of functions to help with input-output analysis, including international and brazilian matrices data, and a convenient addin for easy input.
}
\author{
\strong{Maintainer}: Alberson da Silva Miranda \email{albersonmiranda@hotmail.com} (\href{https://orcid.org/0000-0001-9252-4175}{ORCID})

Other contributors:
\itemize{
  \item Celso Bissoli Sessa \email{celso.sessa@ufes.br} (\href{https://orcid.org/0000-0001-7616-0244}{ORCID}) [data contributor]
  \item Vinícius de Almeida Vale \email{viniciusvale@ufpr.br} (\href{https://orcid.org/0000-0001-5869-9860}{ORCID}) [contributor]
  \item Fernando Salgueiro Perobelli \email{fernando.perobelli@ufjf.edu.br} (\href{https://orcid.org/0000-0003-2364-8865}{ORCID}) [contributor]
}

}
