% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{fio-package}
\alias{fio}
\alias{fio-package}
\title{fio: Friendly Input-Output Analysis}
\description{
A set of functions to perform input-output analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://albersonmiranda.github.io/fio/}
  \item Report bugs at \url{https://github.com/albersonmiranda/fio/issues}
}

}
\author{
\strong{Maintainer}: Alberson da Silva Miranda \email{albersonmiranda@hotmail.com} (\href{https://orcid.org/0000-0001-9252-4175}{ORCID})

Other contributors:
\itemize{
  \item Celso Bissoli Sessa \email{celso.sessa@ufes.br} (\href{https://orcid.org/0000-0001-7616-0244}{ORCID}) [data contributor]
}

}
