% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_iom.R
\name{import_element}
\alias{import_element}
\title{Import IOM data}
\usage{
import_element(file, sheet, range, col_names = FALSE, row_names = FALSE)
}
\arguments{
\item{file}{Path to the Excel file.}

\item{sheet}{Name of the sheet in the Excel file.}

\item{range}{Range of cells in the Excel file.}

\item{col_names}{Range of cells with column names.}

\item{row_names}{Range of cells with row names.}
}
\value{
A matrix with row and column names.
}
\description{
Import data from a input-output matrix (IOM) from Excel format.
}
\examples{
\dontrun{
intermediate_transactions = import_element(
 file = "path/to/file.xlsx",
 sheet = "sheet_name",
 range = "B2:Z56",
 col_names = "B2:Z2",
 row_names = "A2:A56"
)
}
}
