% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{iom}
\alias{iom}
\title{R6 class for input-output matrix}
\description{
R6 class for input-output matrix.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Identifier of the new instance}

\item{\code{intermediate_transactions}}{Intermediate transactions matrix.}

\item{\code{total_production}}{Total production matrix.}

\item{\code{final_demand}}{Final demand matrix.}

\item{\code{exports}}{Exports matrix.}

\item{\code{imports}}{Imports matrix.}

\item{\code{taxes}}{Taxes matrix.}

\item{\code{wages}}{Wages matrix.}

\item{\code{operating_income}}{Operating income matrix.}

\item{\code{added_value_final_demand}}{Value added final demand matrix.}

\item{\code{added_value}}{Value added matrix.}

\item{\code{occupation}}{Occupation vector}

\item{\code{technical_coefficients_matrix}}{Technical coefficients matrix.}

\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-iom-new}{\code{iom$new()}}
\item \href{#method-iom-add}{\code{iom$add()}}
\item \href{#method-iom-remove}{\code{iom$remove()}}
\item \href{#method-iom-tec_coeff}{\code{iom$tec_coeff()}}
\item \href{#method-iom-leontief_inverse}{\code{iom$leontief_inverse()}}
\item \href{#method-iom-clone}{\code{iom$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-new"></a>}}
\if{latex}{\out{\hypertarget{method-iom-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$new(
  id,
  intermediate_transactions,
  total_production,
  final_demand = NULL,
  exports = NULL,
  imports = NULL,
  taxes = NULL,
  wages = NULL,
  operating_income = NULL,
  added_value_final_demand = NULL,
  added_value = NULL,
  occupation = NULL,
  technical_coefficients_matrix = NULL,
  leontief_inverse_matrix = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Identifier for the input-output matrix.}

\item{\code{intermediate_transactions}}{Intermediate transactions matrix.}

\item{\code{total_production}}{Total production matrix.}

\item{\code{final_demand}}{Final demand matrix.}

\item{\code{exports}}{Exports matrix.}

\item{\code{imports}}{Imports matrix.}

\item{\code{taxes}}{Taxes matrix.}

\item{\code{wages}}{Wages matrix.}

\item{\code{operating_income}}{Operating income matrix.}

\item{\code{added_value_final_demand}}{Value added final demand matrix.}

\item{\code{added_value}}{Value added matrix.}

\item{\code{occupation}}{Occupation matrix.}

\item{\code{technical_coefficients_matrix}}{Technical coefficients matrix.}

\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-add"></a>}}
\if{latex}{\out{\hypertarget{method-iom-add}{}}}
\subsection{Method \code{add()}}{
Adds a matrix to a previously imported IO matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$add(matrix_name, matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix_name}}{One of final_demand, exports, imports, added_value_final_demand, added_value or occupation matrix to be added.}

\item{\code{matrix}}{Matrix object to be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-remove"></a>}}
\if{latex}{\out{\hypertarget{method-iom-remove}{}}}
\subsection{Method \code{remove()}}{
Removes a matrix from a previously imported IO matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$remove(matrix_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix_name}}{One of final_demand, exports, imports, added_value_final_demand, added_value or occupation matrix to be removed.}

\item{\code{matrix}}{Matrix object to be removed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-tec_coeff"></a>}}
\if{latex}{\out{\hypertarget{method-iom-tec_coeff}{}}}
\subsection{Method \code{tec_coeff()}}{
Calculate the technical coefficients matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$tec_coeff()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-leontief_inverse"></a>}}
\if{latex}{\out{\hypertarget{method-iom-leontief_inverse}{}}}
\subsection{Method \code{leontief_inverse()}}{
Calculate the Leontief inverse matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$leontief_inverse(technical_coefficients_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{technical_coefficients_matrix}}{Technical coefficients matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-clone"></a>}}
\if{latex}{\out{\hypertarget{method-iom-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
