% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{iom}
\alias{iom}
\title{R6 class for input-output matrix}
\description{
R6 class for input-output matrix.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Identifier of the new instance}

\item{\code{intermediate_transactions}}{Intermediate transactions matrix.}

\item{\code{total_production}}{Total production vector.}

\item{\code{final_demand}}{Final demand matrix.}

\item{\code{exports}}{Exports vector.}

\item{\code{imports}}{Imports vector.}

\item{\code{taxes}}{Taxes vector.}

\item{\code{wages}}{Wages vector.}

\item{\code{operating_income}}{Operating income vector.}

\item{\code{added_value_final_demand}}{Added value final demand matrix.}

\item{\code{added_value}}{Added value vector.}

\item{\code{occupation}}{Occupation vector}

\item{\code{technical_coefficients_matrix}}{Technical coefficients matrix.}

\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}

\item{\code{multiplier_output}}{Output multiplier vector.}

\item{\code{field_influence}}{Influence field matrix.}

\item{\code{key_sectors}}{Key sectors dataframe.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-iom-new}{\code{iom$new()}}
\item \href{#method-iom-add}{\code{iom$add()}}
\item \href{#method-iom-remove}{\code{iom$remove()}}
\item \href{#method-iom-compute_tech_coeff}{\code{iom$compute_tech_coeff()}}
\item \href{#method-iom-compute_leontief_inverse}{\code{iom$compute_leontief_inverse()}}
\item \href{#method-iom-compute_multiplier_output}{\code{iom$compute_multiplier_output()}}
\item \href{#method-iom-compute_field_influence}{\code{iom$compute_field_influence()}}
\item \href{#method-iom-compute_key_sectors}{\code{iom$compute_key_sectors()}}
\item \href{#method-iom-clone}{\code{iom$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-new"></a>}}
\if{latex}{\out{\hypertarget{method-iom-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$new(
  id,
  intermediate_transactions,
  total_production,
  final_demand = NULL,
  exports = NULL,
  imports = NULL,
  taxes = NULL,
  wages = NULL,
  operating_income = NULL,
  added_value_final_demand = NULL,
  added_value = NULL,
  occupation = NULL,
  technical_coefficients_matrix = NULL,
  leontief_inverse_matrix = NULL,
  multiplier_output = NULL,
  field_influence = NULL,
  key_sectors = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Identifier for the input-output matrix.}

\item{\code{intermediate_transactions}}{Intermediate transactions matrix.}

\item{\code{total_production}}{Total production vector.}

\item{\code{final_demand}}{Final demand matrix.}

\item{\code{exports}}{Exports vector.}

\item{\code{imports}}{Imports vector.}

\item{\code{taxes}}{Taxes vector.}

\item{\code{wages}}{Wages vector.}

\item{\code{operating_income}}{Operating income vector.}

\item{\code{added_value_final_demand}}{Added value final demand matrix.}

\item{\code{added_value}}{Added value vector.}

\item{\code{occupation}}{Occupation matrix.}

\item{\code{technical_coefficients_matrix}}{Technical coefficients matrix.}

\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}

\item{\code{multiplier_output}}{Output multipler vector.}

\item{\code{field_influence}}{Influence field matrix.}

\item{\code{key_sectors}}{Key sectors dataframe.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-add"></a>}}
\if{latex}{\out{\hypertarget{method-iom-add}{}}}
\subsection{Method \code{add()}}{
Adds a matrix to a previously imported IO matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$add(matrix_name, matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix_name}}{One of final_demand, exports, imports, added_value_final_demand, added_value or occupation matrix to be added.}

\item{\code{matrix}}{Matrix object to be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-remove"></a>}}
\if{latex}{\out{\hypertarget{method-iom-remove}{}}}
\subsection{Method \code{remove()}}{
Removes a matrix from a previously imported IO matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$remove(matrix_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix_name}}{One of final_demand, exports, imports, added_value_final_demand, added_value or occupation matrix to be removed.}

\item{\code{matrix}}{Matrix object to be removed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_tech_coeff"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_tech_coeff}{}}}
\subsection{Method \code{compute_tech_coeff()}}{
Computes the technical coefficients matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_tech_coeff()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_leontief_inverse"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_leontief_inverse}{}}}
\subsection{Method \code{compute_leontief_inverse()}}{
Computes the Leontief inverse matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_leontief_inverse(technical_coefficients_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{technical_coefficients_matrix}}{Technical coefficients matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_multiplier_output"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_multiplier_output}{}}}
\subsection{Method \code{compute_multiplier_output()}}{
Computes the output multiplier vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_multiplier_output(leontief_inverse_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_field_influence"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_field_influence}{}}}
\subsection{Method \code{compute_field_influence()}}{
Computes the influence field matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_field_influence(epsilon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epsilon}}{Epsilon value. A technical change in the input-output matrix.}

\item{\code{technical_coefficients_matrix}}{Technical coefficients matrix.}

\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_key_sectors"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_key_sectors}{}}}
\subsection{Method \code{compute_key_sectors()}}{
Computes the key sectors dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_key_sectors(leontief_inverse_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-clone"></a>}}
\if{latex}{\out{\hypertarget{method-iom-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
