% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6.R
\name{iom}
\alias{iom}
\title{R6 class for input-output matrix}
\description{
R6 class for input-output matrix.
}
\examples{
# running single threaded to comply with CRAN policies. Ignore for performance.
fio:::set_max_threads(1L)
# data
intermediate_transactions <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), 3, 3)
total_production <- matrix(c(100, 200, 300), 1, 3)
exports <- matrix(c(10, 20, 30), 3, 1)
households <- matrix(4:6, 3, 1)
imports <- matrix(c(5, 10, 15), 1, 3)
jobs <- matrix(c(10, 12, 15), 1, 3)
taxes <- matrix(c(2, 5, 10), 1, 3)
wages <- matrix(c(11, 12, 13), 1, 3)

# a new iom instance can be created by passing just intermediate transactions and total production
my_iom <- iom$new(
 "example",
 intermediate_transactions,
 total_production
)

# Compute technical coefficients matrix
my_iom$compute_tech_coeff()
print(my_iom$technical_coefficients_matrix)

# Compute leontief inverse matrix
my_iom$compute_leontief_inverse()
print(my_iom$leontief_inverse_matrix)

# compute output multiplier
my_iom$compute_multiplier_output()
print(my_iom$multiplier_output)

# `add` or `remove` other elements to IO matrix
my_iom$add("exports", exports)
my_iom$add("household_consumption", households)
my_iom$add("occupation", jobs)
my_iom$add("taxes", taxes)
my_iom$add("wages", wages)

# compute multipliers for added elements
my_iom$compute_multiplier_employment()
my_iom$compute_multiplier_taxes()
my_iom$compute_multiplier_wages()
print(my_iom$multiplier_wages)

# compute field of influence
my_iom$compute_field_influence(epsilon = 0.001)

# compute power and sensitivity of dispersion, it's coefficients of variation
# and identify key sectors
my_iom$compute_key_sectors()
print(my_iom$key_sectors)

# supply-wise model
my_iom$compute_allocation_coeff()
my_iom$compute_ghosh_inverse()
print(my_iom$ghosh_inverse_matrix)

# aggregates final demand and added value vectors
my_iom$update_final_demand_matrix()
print(my_iom$final_demand_matrix)

my_iom$update_added_value_matrix()
print(my_iom$added_value_matrix)

# perform hypothetical extraction of a given sector
my_iom$compute_hypothetical_extraction()
print(my_iom$hypothetical_extraction)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Identifier of the new instance}

\item{\code{intermediate_transactions}}{Intermediate transactions matrix.}

\item{\code{total_production}}{Total production vector.}

\item{\code{household_consumption}}{Household consumption vector.}

\item{\code{government_consumption}}{Government consumption vector.}

\item{\code{exports}}{Exports vector.}

\item{\code{final_demand_others}}{Other vectors of final demand that doesn't have dedicated slots.}

\item{\code{final_demand_matrix}}{Aggregates final demand vectors into a matrix.}

\item{\code{imports}}{Imports vector.}

\item{\code{taxes}}{Taxes vector.}

\item{\code{wages}}{Wages vector.}

\item{\code{operating_income}}{Operating income vector.}

\item{\code{added_value_others}}{Other vectors of added value that doesn't have dedicated slots.}

\item{\code{added_value_matrix}}{Aggregates added value vectors into a matrix.}

\item{\code{occupation}}{Occupation vector}

\item{\code{technical_coefficients_matrix}}{Technical coefficients matrix.}

\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}

\item{\code{multiplier_output}}{Output multiplier dataframe.}

\item{\code{multiplier_employment}}{Employment multiplier dataframe.}

\item{\code{multiplier_taxes}}{Taxes multiplier dataframe.}

\item{\code{multiplier_wages}}{Wages multiplier dataframe.}

\item{\code{field_influence}}{Influence field matrix.}

\item{\code{key_sectors}}{Key sectors dataframe.}

\item{\code{allocation_coefficients_matrix}}{Allocation coefficients matrix.}

\item{\code{ghosh_inverse_matrix}}{Ghosh inverse matrix.}

\item{\code{hypothetical_extraction}}{Absolute and relative backward and forward differences in total output after a hypothetical extraction.}

\item{\code{threads}}{Number of threads available for Rust to run in parallel.
Defaults to 0, meaning running in parallel with all available threads.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-iom-new}{\code{iom$new()}}
\item \href{#method-iom-add}{\code{iom$add()}}
\item \href{#method-iom-remove}{\code{iom$remove()}}
\item \href{#method-iom-update_final_demand_matrix}{\code{iom$update_final_demand_matrix()}}
\item \href{#method-iom-update_added_value_matrix}{\code{iom$update_added_value_matrix()}}
\item \href{#method-iom-compute_tech_coeff}{\code{iom$compute_tech_coeff()}}
\item \href{#method-iom-compute_leontief_inverse}{\code{iom$compute_leontief_inverse()}}
\item \href{#method-iom-compute_multiplier_output}{\code{iom$compute_multiplier_output()}}
\item \href{#method-iom-compute_multiplier_employment}{\code{iom$compute_multiplier_employment()}}
\item \href{#method-iom-compute_multiplier_wages}{\code{iom$compute_multiplier_wages()}}
\item \href{#method-iom-compute_multiplier_taxes}{\code{iom$compute_multiplier_taxes()}}
\item \href{#method-iom-compute_field_influence}{\code{iom$compute_field_influence()}}
\item \href{#method-iom-compute_key_sectors}{\code{iom$compute_key_sectors()}}
\item \href{#method-iom-compute_allocation_coeff}{\code{iom$compute_allocation_coeff()}}
\item \href{#method-iom-compute_ghosh_inverse}{\code{iom$compute_ghosh_inverse()}}
\item \href{#method-iom-compute_hypothetical_extraction}{\code{iom$compute_hypothetical_extraction()}}
\item \href{#method-iom-set_max_threads}{\code{iom$set_max_threads()}}
\item \href{#method-iom-clone}{\code{iom$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-new"></a>}}
\if{latex}{\out{\hypertarget{method-iom-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$new(
  id,
  intermediate_transactions,
  total_production,
  household_consumption = NULL,
  government_consumption = NULL,
  exports = NULL,
  final_demand_others = NULL,
  imports = NULL,
  taxes = NULL,
  wages = NULL,
  operating_income = NULL,
  added_value_others = NULL,
  occupation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Identifier for the input-output matrix.}

\item{\code{intermediate_transactions}}{Intermediate transactions matrix.}

\item{\code{total_production}}{Total production vector.}

\item{\code{household_consumption}}{Household consumption vector.}

\item{\code{government_consumption}}{Government consumption vector.}

\item{\code{exports}}{Exports vector.}

\item{\code{final_demand_others}}{Other vectors of final demand that doesn't have dedicated slots.
Setting column names is advised for better readability.}

\item{\code{imports}}{Imports vector.}

\item{\code{taxes}}{Taxes vector.}

\item{\code{wages}}{Wages vector.}

\item{\code{operating_income}}{Operating income vector.}

\item{\code{added_value_others}}{Other vectors of added value that doesn't have dedicated slots.
Setting row names is advised for better readability.}

\item{\code{occupation}}{Occupation matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-add"></a>}}
\if{latex}{\out{\hypertarget{method-iom-add}{}}}
\subsection{Method \code{add()}}{
Adds a matrix to a previously imported IO matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$add(matrix_name, matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix_name}}{One of household_consumption, government_consumption, exports, final_demand_others,
imports, taxes, wages, operating income, added_value_others or occupation matrix to be added.}

\item{\code{matrix}}{Matrix object to be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-remove"></a>}}
\if{latex}{\out{\hypertarget{method-iom-remove}{}}}
\subsection{Method \code{remove()}}{
Removes a matrix from a previously imported IO matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$remove(matrix_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix_name}}{One of household_consumption, government_consumption, exports, final_demand_others,
imports, taxes, wages, operating income, added_value_others or occupation matrix to be removed.}

\item{\code{matrix}}{Matrix object to be removed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-update_final_demand_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-iom-update_final_demand_matrix}{}}}
\subsection{Method \code{update_final_demand_matrix()}}{
Updates final demand matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$update_final_demand_matrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-update_added_value_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-iom-update_added_value_matrix}{}}}
\subsection{Method \code{update_added_value_matrix()}}{
Updates added value matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$update_added_value_matrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_tech_coeff"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_tech_coeff}{}}}
\subsection{Method \code{compute_tech_coeff()}}{
Computes the technical coefficients matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_tech_coeff()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_leontief_inverse"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_leontief_inverse}{}}}
\subsection{Method \code{compute_leontief_inverse()}}{
Computes the Leontief inverse matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_leontief_inverse()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{technical_coefficients_matrix}}{Technical coefficients matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_multiplier_output"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_multiplier_output}{}}}
\subsection{Method \code{compute_multiplier_output()}}{
Computes the output multiplier dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_multiplier_output()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_multiplier_employment"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_multiplier_employment}{}}}
\subsection{Method \code{compute_multiplier_employment()}}{
Computes the employment multiplier dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_multiplier_employment()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_multiplier_wages"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_multiplier_wages}{}}}
\subsection{Method \code{compute_multiplier_wages()}}{
Computes the wages multiplier dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_multiplier_wages()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_multiplier_taxes"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_multiplier_taxes}{}}}
\subsection{Method \code{compute_multiplier_taxes()}}{
Computes the taxes multiplier dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_multiplier_taxes()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leontief_inverse_matrix}}{Leontief inverse matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_field_influence"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_field_influence}{}}}
\subsection{Method \code{compute_field_influence()}}{
Computes the influence field matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_field_influence(epsilon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epsilon}}{Epsilon value. A technical change in the input-output matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_key_sectors"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_key_sectors}{}}}
\subsection{Method \code{compute_key_sectors()}}{
Computes the key sectors dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_key_sectors()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_allocation_coeff"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_allocation_coeff}{}}}
\subsection{Method \code{compute_allocation_coeff()}}{
Computes the allocation coefficients matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_allocation_coeff()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_ghosh_inverse"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_ghosh_inverse}{}}}
\subsection{Method \code{compute_ghosh_inverse()}}{
Computes the Ghosh inverse matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_ghosh_inverse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-compute_hypothetical_extraction"></a>}}
\if{latex}{\out{\hypertarget{method-iom-compute_hypothetical_extraction}{}}}
\subsection{Method \code{compute_hypothetical_extraction()}}{
Computes the hypothetical extraction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$compute_hypothetical_extraction()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-set_max_threads"></a>}}
\if{latex}{\out{\hypertarget{method-iom-set_max_threads}{}}}
\subsection{Method \code{set_max_threads()}}{
Sets max number of threads used by fio.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$set_max_threads(max_threads)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_threads}}{Number of threads enabled for parallel computing. Defaults to the number
of threads available.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iom-clone"></a>}}
\if{latex}{\out{\hypertarget{method-iom-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
